Fix computing forces with FAR model when model was not ready
Computing forces with the FAR model requires that the active vessel is not packed and that its FARVesselAero component reports that has a valid voxelization. If the forces were computed before that, they would be zero. The cache would then be populated with these zero forces and the trajectory would be wrong.

This fix introduces a new method IsReady() for AeroDynamicModel, which reports if the aerodynamic model is ready to compute forces. Forces are only computed when this method returns true.